+++ src/test.utils.ts
@@ -1,3 +1,8 @@
+import { expect } from "vitest";
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Constructor<T extends object = object> = new (...args: any[]) => T;
+
 export async function aggregateAsyncIterable<TYield, TReturn>(
 	iterable: AsyncIterable<TYield, TReturn>,
 ) {
@@ -25,3 +30,23 @@ export async function aggregateAsyncIterable<TYield, TReturn>(
 		};
 	}
 }
+
+export async function waitError<TError extends Error = Error>(
+	fnOrPromise: (() => unknown) | Promise<unknown>,
+	errorConstructor?: Constructor<TError>,
+): Promise<TError> {
+	try {
+		if (typeof fnOrPromise === "function") {
+			await fnOrPromise();
+		} else {
+			await fnOrPromise;
+		}
+	} catch (cause) {
+		expect(cause).toBeInstanceOf(errorConstructor ?? Error);
+		return cause as TError;
+	}
+	throw new Error("Function did not throw");
+}
+
+export const sleep = (ms: number) =>
+	new Promise((resolve) => setTimeout(resolve, ms));
